chore(root): adds loose node 20 support
